Added an abstract scene descriptor to the place object.
The goal is to abstract away how the scene is transmitted over the wire. In some cases, maybe we want to send 
the full scene. In others, maybe we only want to send an idea and have the client load a level locally. In 
yet others, maybe we send a random seed and some parameters and have the client regenerate a random level. At 
least that's the goal.


git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@115 edc67d0a-ffea-dc11-983e-003048321218