Let's use a CopyOnWriteArrayList.
It's ideally suited to notifying listeners because any modifications
to the list are not seen during iteration.
We create a small amount of garbage (an Iterator) now, but this is
much safer and better than always copying the elements out to an
array during dispatch (even if re-using the array).


git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@1621 edc67d0a-ffea-dc11-983e-003048321218