Added a scrollRectToVisible().
Something in here is not quite right. I'm getting flummoxed by insets or
something else, but this is pretty much working where I need it to work
(there's a tiny bit of ugly). Rather than continue to bang on it, I'm
going to put a pin in it and we can revisit it later.


git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@1693 edc67d0a-ffea-dc11-983e-003048321218