Aiya, so when the TextFactory breaks lines it removes trailing whitespace or any newline that forced the break. Previously I was just dealing with it by adding 1 to the length of every line.
But as they say: There are 2 common kinds of bugs: off by 1 errors.

The +1 addition broke if the line wrapped on a non-whitespace boundary.
So: when turning a mouse position into a cursor position, or a cursor
position into a screen position, we need to see what's actually in the document
at the end of the line and add 1 if it's whitespace. Hopefully there
are no more surprises there.

Also: only scroll the cursor to be visible if we're actually displaying it.


git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@1696 edc67d0a-ffea-dc11-983e-003048321218