Bugfix: noticed that in the scene editor I could sometimes click a tile and then click another tile that had the same config and the editor would show the values for the first.
It turns out to merely have been a small gui bug, it just
needed refreshing.

Although, things look a little off in here, if an editor is reused
and removed from the JPanel, then the ocount is not updated.
Probably it'd be safest to just remove all the old ones and re-add
them as needed.


git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@1754 edc67d0a-ffea-dc11-983e-003048321218