Andrzej, please take a look at this when you're back.
So using timestamps to handle the id's of prespawned sprites is problematic.  I've found 3 issues
so far.

1) We sometimes have the client side try to do a prespawn twice on the same timestamp.  This seems
to be a duplicate call somehow, I didn't look into it too deeply, but doesn't seem to cause any
real problems other than a second actorsprite replacing an existing one at the id so the old one
is not being properly removed.  I haven't added anything to resolve this issue.

2) We prespawn a sprite with a timestamp and somehow the server will send a tick with that
timestamp without having created the actor.  So the client removes the prespawn sprite, and the
next tick it will get the servers actor created, and not find the sprite.

3) The client and server don't agree on the created timestamp for the actor.  In my testing this
seems to always involved the client thinking it's ahead of the server.  In any case, since the
timestamps don't match up, the prespawn sprite is removed, and the server version never gets an
associated sprite.

To solve both 2 & 3, I've changed TudeySceneView to create a new actor sprite if it was unable to
find the prespawned one.  I can potentially see issues with this where sprites may show up twice
or reset depending on how the prespawn sprite will act (I'm thinking of how BulletSprite operates
differently for the local client than the server).  But in any of these cases, the prespawned sprite
should have already been removed, so there should still only exist one in the view at a time.  And
it's better than the issue of not seeing what you should be. (Ie: your bomb disappearing).


git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@1877 edc67d0a-ffea-dc11-983e-003048321218