- Change the translation used for a null value to be less likely   to interfere with a valid translation for something else.
So... having null translate as m.none interferes with the ability to
have an enum value called NONE.

The real problem here is JComboBox, which just toStrings() its values
to display them, so we typically provide it with a String[] when we
want to customize the names. Two m.none's are coming in and it doesn't
distinguish between them.

Ideally a ComboBox (or anything list-like) would take the value Objects
and some interface for formatting those into a String.
(The opengl.gui.ComboBox has a Formatter interface for this.)

JComboBox has a ListCellRenderer, but that has a bunch of extra stuff.
I could still wire one up that takes a Function for formatting and
passes the rest through to the default renderer. But then there's
the ComboBoxEditor, etc, etc. Plus we don't want this to be cumbersome
to set up. So for now I just did this.


git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@1908 edc67d0a-ffea-dc11-983e-003048321218