Changed the Controller interface to have a single setActive() method instead of start() and stop(). This method is expected to be idempotent- the UI isn't going to keep track of whether it has already informed the controller of active/inactive-ness, it's just going to call this when it may have changed.
Clyde components have an isShowing() method, but it doesn't work
the same way as AWT/Swing, which are only "showing" if they are visible,
added to the hierarchy, and *so are all the parents, up the hierarchy*.
In Clyde, isShowing() may return true even if the parent is invisible.

Maybe I'll fix that if it doesn't break anything.

For now, its sufficient for my needs to tie the controller's active-ness to
isAdded() && isEnabled() && isVisible().


git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@1964 edc67d0a-ffea-dc11-983e-003048321218