I was annoyed that MenuItems weren't disarmable. Once you press the mouse down on them, they'll fire on mouseup even if you've since clicked button2 or moved the mouse out of their bounds.
Button had proper behavior. So I pushed that logic into a new
AbstractButton that both Button and MenuItem extend. I had to
rejigger some things so that the logic worked for both of them,
and I can give feedback as the button arms and disarms even
though MenuItems don't have a DOWN state.

There was some code in Button to detect if the mouse was released
at the same time as a drag that Mark added in 2009. I adapted that,
but while testing I couldn't trigger it at all.

After numerous attempts to discern what it was doing, I gave up and nixed
the code, which simplified things even more. As far as I can tell,
everything is behaving ideally.

It's possible that it's a platform difference, and I've now made
buttons annoying on Windows or Linux. I guess I'll just have to check
this in and we'll test it out. It should be easy for me to fix.


git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@1973 edc67d0a-ffea-dc11-983e-003048321218