Add new property editors: datetime, date, & time. Can edit Date, Long, or long. Nullable.
In terms of usability it's only a minor improvement because the
date is still edited as a String, but it will parse it during
edit time: highlighting in red if the text is invalid and
confirming the edit by re-formatting the date when focus is lost.

Behind the scenes it's a bigger improvement because we won't be
subject to timezone or locale formatting assumptions (aka bugs),
and we shouldn't be parsing dates in logic classes anyway.

The "mode" attribute can be used to specify the DateFormat style,
the Locale, and the TimeZone.


git-svn-id: svn+ssh://src.earth.threerings.net/clyde/trunk@2001 edc67d0a-ffea-dc11-983e-003048321218