An improved fix to model visibility.
We don't want to unilaterally remove the implementation from
the scene when visibility changes: most of the time the
visibility is effected by not compositing the implementation.
Added setVisible() on Model.Implementation, which is mostly a no-op
except that SceneInfluencers and ViewerAffecters will remove
their influence/effect, and a few other types will simply relay
the call to their contained implementations.

This should allow me to make most kinds of GlobalSprite invisible.