Update for animations (especially procedural) to be correct.
1.) Add method to DynamicScope to add a listener to an index in the listener queue.

2.) Add procedural animations to the front of the listener queue,
thereby allowing them to be executed last. This is because procedural
animations require the use of their parents' timestamps.

3.) Add method to scopeUtil to resolve timestamps, allowing for lazy use of
System.currentTimeMillis instead of constant use as an argument.

4.) Switch everything checking for a timestamp over to the new method.

5.) Fix animations to rely on internal timestamping instead of external 'elapsed'
because they may be removed from tickable queue at any time (thus making elapsed invalid).
This change, or something like it, may be brought up to the model level.