Fix animation off-by-one tick error.
Animations now use a timestamp measured within the current frame instead
of the elapsed variable. Because elapsed is measured at the end of a frame
and the timestamps within, an off-by-one issue arose that lead to some
glitchy animations.

This is likely not the best solution to this issue, and may be addressed
further in the future.