Added convenience methods for getting ComboBoxes.
This avoids callers needing to SuppressWarnings and know
when/how it is safe to do so.

There are two methods here, because if you get a ComboBox<String>
you may want to save the values configured in the UI.

But if it's not of type String then it's only safe if we clear it.