Preserve selection when filter changes; clear filter on "new".
So, things are a little weird because:
- There is not a good way to filter a tree on the "view" side,
  instead the model must be filtered.
- The ConfigTree stores a lot of "editor state" in the model
  and nodes.

Because of that, it's not a good idea to add new configs when in
a filtered view. If you create a new config with a name that's
already in use, it won't realize it and will overwrite the
original config. If you create a folder, but add no configs,
and then change your filtering, the folder will be gone.

So: clear out the filter prior to adding a new node.

Also: preserve selection on filter changes. If you've filtered
such that your original selection is not visible, but don't
select anything new in the filtered view, when you remove
the filter you'll find your original selection still selected.

I can probably do a larger refactor and possibly fix some of this.
The most obvious thing would be to simply do some overrides
in ConfigTreeNode to make it "hide" children or siblings that
don't match the predicate, but they're still actually there.

I may try that next.