Modify Scripted to support/obey a duration constant.
-Modified ScriptedConfig to include a duration constant.
-Modified Scripted to utilize a timestamp instead of a sumation of elapsed time.
This causes Scripted to be more accurate with its understanding of time, and thus
allows us to use the above duration constant to judge whether an effect's
length has already been surpased. If so, we may skip it.

This should theoretically have no functional effect on Scripted behavior, but
does mean effects may fire slightly differently in situations where ticking
was only being done on-screen. The intent is to follow this change with a debug
message to help weed out those situations.