Patch LWJGL's improper handling of the mac ` (grave) key.
Mac's keyboard under LWJGL is handled through a mapping of AWT to
LWJGL keys. Unfortunately, grave is not mapped at all (let alone
correctly). We still do, however, receive the event character
correctly, so we can use that to fix the broken event key.
LWJGL should be moving to native library for Mac, so in the future
this patch should be removed.