Add initialization action for UIs.
This change adds an 'Init Action' to the UI. This allows initial states
to be setup, especially with respect to alpha, offset and visibility,
so that animations will not have a frame (or more) of inaccurate
playback on screen prior to their initial run. (For instance, an animation
of a UI sliding in from the left from off screen was previously somewhat
ruined because the UI would first show up on screen before being moved.)

The init action has several limitations, chief among them is that it is
not ticked and therefore cannot playback any animations. Attempting to
do so will likely yield error messages. However, setX (e.g. setVisible,
setAlpha, setOffset, etc.) actions work.