Redefine isShowing() to include all parents.
isShowing() will return true when a component is added to the
hierarchy, visible, and all parents up the hierarchy are also visible.

This is potentially a breaking change, but I'm checking the uses of
isShowing here and on px to ensure that everything's happy. All here
in clyde should be fine.