Fake a mouse exit on the clicked Component when dragging starts.
When a drag starts, the originally clicked component is cleared and so
will no longer receive mouse events. If it was a button, it may have been
armed and will not properly learn about the mouse release that eventually
happens.

So: fake up some mouse events so that the component knows that the mouse
has left.

Apparently this wasn't noticed before because the affected buttons used
the same style for the down and up states... :O