Efficiency.
How many times shall we test for null?
Why retrieve the annotation (which never throws) when it's
only checked if the value is null?

Also: remove ancient TODO. That may be why the annotation was
retrieved... so that the value could be validated against
other constraints of the annotation. But since that's never been
done I'm going to assume we're fine with not starting now.