Efficiency.
How many times shall we do a map lookup?
Previously it was 1 for a missing value, 2 for an invalid value,
and a full 3 times for a valid value.
Let's cut that down to 1, or 2 if the value is null or missing.
Not benchmarked, but I would think that having a value is
the common case.