Add 'completed actions' to moveOffset and fadeAlpha.
Added a completed actions field to moveOffset and fadeAlpha. This
allows explicit continuation from these (under the hood) interface
scripts, and helps prevent issues where during cleanup they might
set a value after the parent script during the same tick. I.e.:

Parent script contains -
time 0.0, move offset, interval 0.25f
time 0.26, set offset 0.

The final step for the move offset is NOT guaranteed to execute prior
to the set offset command at 0.26s, despite the fact that the move
offset is supposed to be shorter. This is because the move offset
command actually adds a script to the interface which, if the two
both expire simultaneously, will tick AFTER the original, executing
script. Because there is no way to guarantee ticks, prior to this
there was no way to guarantee execution of elements after animation
completion.

NOTE: Wait actions do not currently function within these new sets
of actions. To get around this, run a script.