Change ColorPicker to use sizing properly to figure out swatch dimensions.
ColorPicker previously used a manual setting for the swatch sizing instead of using the
preferred dimensions and simply setting up the swatches based upon that. Since we're
now using multiple swatches, it's pretty useful to have it automate this process instead
of having to mess with the UI elements all the time.

Also added a method to get the string given for the color class.