Add missing SceneInfluencer and ViewerAffecter check for being added
SceneInfluencer and ViewerAffecter needed to check that they were
added before going through with updating their bounds. This is because
the operation of updating bounds causes their respective subelements
to be removed and added to the scene. Double-adding or double-removing
from the scene appears to cause a lingering-element bug (for presently
unknown reasons that weren't worth investigating).

Also, only send down the visbilityWasSet signal when there was a change
in visibility (otherwise the requisite upwards traversal will be done
every tick on some props/sprites).