Use a Closer instead of the deprecated Closeables.closeQuietly()
This may not be as "quiet", but that's part of the point.
Closing quietly is fine for an input stream, but for an output
stream an IOException on closing might be the first time that a
write is making its way through various buffers and actually hitting
the IO system.

The usage pattern for Closer is a strange duck indeed.