Support qualified message keys in UIs.
It'd be great if we could just xlate everything, and then we could
use qualified keys, composed keys, or (rarely) taint things that
are supposed to be used as-is.

We can't do that now, because we have piles of legacy UIs that would
need updating, but I can move towards this by supporting tainting
now and attempting to determine when we can safely pass to xlate.