Bind some config editor UI properties to preferences.
I gotta believe there's a library somewhere that makes this less painful,
but given that some of these are bean bound properties while some are
something else, and that we're saving to Preferences in the end which
have their own legacy issues, maybe the pain is fairly unavoidable.

This might need tweaking, but at least on my OS when opening multiple
copies of the config editor the window positions cascade nicely.