Experimental fixup for bogus mouse clicks.
Previously I ignored the mouse click that focused the window, because it
would not have the right coordinate. However, it still has problems
because even the next click may not have the correct mouse position.
It seems that until a mouse move event is generated, LWJGL has a cached
old position of the mouse coordinate, and even polling it during
the focusing click will not get the right value.

So it was possible to click in the window and have a gui element other
the one you're hovering over receive the click, and it sucked.

So now I ignore mouse clicks until we've received a mouse move event.
It's a little bit weird that you can focus the window and click away
without doing anything, but I feel like it's much better than the
old behavior of clicking on the wrong thing...