Experiment, Mark II.
Use the AWT to locate the correct position of the mouse.
Use that while processing events during a new focus, and then
force LWJGL to reposition the mouse to that location for the next update.

This seems to work remarkably well. I'm loving it, except for the 1-pixel
fudge factor that seemed necessary with the Y coordinate...