Copy a HashIntMap by cloning it.
I was looking for a convenient copying factory earlier. I wanted
to add one but didn't want to muck in samskivert. But a careful
look at HashIntMap revealed that it's fully cloneable and so that
is almost certainly the most efficient way to make a copy.

Mind you, I'm not sure I "believe" in using the primitive collections
anymore, but I'm not going to tweak the fabric of tudey without
doing some benchmarking.