Rethink the defaults for @Editable's width & height.
Previously the defaults were 10 and 1 respectively. The problem
was that the different editors interpreted these values slightly
differently and for some a width of 10 is too small. They needed
a larger default but also a way to know when the user has actually
specified a width in the annotation so that their default could
be overridden when necessary.

I find myself talking about "default defaults" a lot lately.
Each editor needs a different default, so a default default that
was a valid value made their job harder.

So now the defaults are 0 and 0. Each editor has its own ideas as
to what its default width and height are (if even applicable) and
now they can use the convenience methods on Property to get any
user-defined value if specified or fall back to their defaults.

This also removes a hack with NumberEditor which would ignore
any specified width unless mode was "sized".