Allow the mode attribute to specify... a mode.
It's confusing, but now you can say mode="mode=time" to get the time editor.
This is to support using the DateRange object with mode=date, if desired.

Although, I just realized this will screw with the 'about now'
calculation it does.
Sheeeit. Maybe the editor itself should be handling the conversion from 0
that we frequently do when editing dates...