Native exporting of Multiset.
Will export the entries and counts, rather than dumbly treating
it as a collection and exporting each element 'count' times.
Old exported Multisets will import correctly.
I left in some pieces related to Multimaps too, but those are not
yet implemented.

Of course, the rabbit hole goes deeper. This just solves a minor
inefficiency, but the real problem we're having is that the DeepUtil
stuff is choking when trying to clone a Multiset.
Unbelievably, it doesn't seem to have any special Collection handling,
instead relying purely on reflection.
It chokes trying to copy a HashMultiset because it can't instantiate
HashMap$Entry... I'm not sure what's going on here, because AFAIK
it can copy a HashMap...

I'm probably not going to wade into those waters...