Stop resetting the start/stop times during editing.
First you set up your start time as 2014-04-01 12:00.
The stop time advanced to match. Then, on the stop time, you want to
edit the month from "04" to "11" but when you type the first "1" it's
intantly parsed as January and the start time is reset.

It happens with years as well and I considered just not accepting
years before 2000, but the month problem is unfixable.

I could make it so that you have to press enter in the editor box but
ain't nobody got time for that.

Added an isValid() method that checks if the start/stop times are
crossed. Not yet used.