Reinstated MacOSX 'grave' key handling.
My current feeling is that MacOSX is doin' it right and the other
platforms are the weird ones- reporting the keycode as grave (~)
when you push backtick (`). But we have a history of switching
the Mac to match the others, and we have default keybindings
and saved user preferences. There's no need to rock this boat.

The remapping works differently now. By default LWJGL uses "native"
keyboard handling on MacOS now and we have to adjust the keymap
differently, but if the native keyboard handler is null I fall
back to hacking the mapping simiarly to the way it previously did.