Remove the MacOSX focusing-click mouse-move hack.
It no longer works with the new LWJGL as the main game frame is
no longer an AWT window.

So: the issue now returns that this was trying to solve:
If you refocus the window with a click; don't move the mouse
a single pixel; then click again, the click location will be
at the place where the mouse was when the window originally lost focus.

I'm leaving the hack in here, commented out, while I look for another fix.