Added TudeySceneModel.addObserverLast()
I need the layer tool to be the last observer so that
it can hear about a new entry after the entry's layer
has been set. I started to do some new OrderedObserver
interface, but reordering the existing observers is a pain
and so while this is a little funny it's much much much less of a hack.