Fire a property change when the entire object changes.
Jeez.
Maybe, since apparently everything existing seems to work
fine without this, I can just do this in the path editor
or something...
But c'mon: if the entire object changes that's the epitome of
a notification-worthy change. Maybe the manual update() calling
when a new object is set at the top level would be unnecessary...