Effectively revert the change to DeepObject.toString() from a few months ago.
Most of the "deep" stuff was developed without this, and I added it purely
so that DateRange toString'd the way I wanted. Well, it prints out
WAY too much detail on things like a ConfigReference, and dives way too
deeply, and is susceptible to loops in the object graph.

Go back to the old way. Leave the toString (Object obj) method in DeepUtil
in case anybody wants to use it.