Checkpoint: editing a List or array of config references directly.
This code is the result of much mashing about, and finally getting
it to work. There is other code in the editors that seems to want
to edit arrays of config refs, and even now I'm not sure why they
don't work. But this works, and it is subtly different than the
other bits.