Tweak the flattener a bit.
Prepare for an ant task.
Also: a different way of flattening derived configs that
should be identical under currently-committed code.

I'm in the middle of working on DerivedConfig, hopefully fixing
up some of the loose ends. So that part is related to that, because
the "getActualConfig" method will hopefully go away.