Hmm.
If the derived raw config has a Choice parameter, I can adapt
that to a direct parameter in the base.
But, the other way around is a mess... if derived raw has a Direct
with one path, it can map that to the Choice in the base, but almost
every other combination is crazy.

Wait- why do I need this? Do we really need to translate the parameters?
Perhaps that's just an edge case with derived config.
Note: applying parameters works 100%. The issue is whether you see
parameters that "make sense" on the config that was generated when
you request a derived config.