Fail validation if there are null refs in a List.
The ship has sailed on regular refs long ago.

See- this is what I'm talking about.
If you have an @Editable(nullable=false) and it's null, then
the property that represents that value should KNOW that, and
know that null is not kosher, and immediately say "this is invalid".
Not the property, but the value represented by the property.

Instead it's all special casey.
The validator gives null-valued properties a pass- they aren't even
examined further. It doesn't care. So each editor has to explicitly
and by itself check for null and decide whether to inform you
by coloring the editor for that property red.

It's waaaaaay too late to fix this, unless we wanted to introduce
a compatibility mode for old projects.