Move toSortedArray() to ConfigManager.
Nothing to see here. Move along.
Basically: it would be nice to be able to hook into this
saving process but subclassing ConfigGroup is a nonstarter.
We've subclassed the manager a few times for different needs
and so let's have the config manager marshall the configs
up for saving and maybe it can do some fun stuff..