Display null config refs as invalid.
Well, this is another difference between editor and validation.

I tried to do it in the ConfigReferencePanelArrayList editor, as the
validator will complain about nulls in a list like that, but..
nonstarter.

If the ConfigReference is in a field, plain, then the "nullable"
property of the annotation is checked, otherwise null is always red.