Also ok to be nullable if it comes from a parameter.
Sigh.
I'd blacklist the bad ones but they're a protected class.

If this causes any more trouble I'll revamp how nullable() works
and make it a method on Property that by default checks the annotation,
but that could be overridden. Then, I'd convert alllllll the places
that currently check the annotation...