PropertyTransformerTask
I was thinking I'd check in alternate i18n files so that
we could see exactly what was translated and what was not...
Maybe swedish chef-inate them.

But it would be even easier to just have a debug mode in
MessageBundle, so that all returned text is prepended with an "x",
or rot13'd, or something. Derp.