Be smarter with @Editable List properties.
I was going to add support for editing an ImmutableList.
But ImmutableList doesn't support null elements, and when
inserting a new element they are quite often null initially.

Instead, let's let the default prototype value be an ImmutableList
(or Collections.emptyList()), and when you add an element it will
immediately convert it to an ArrayList.

Note: you can use a pre-populated ImmutableList.

Similarly, when you remove the last element in the list, let's
recheck the value from the prototype, and if it's null,
Collections.emptyList() or ImmutableList.of() (only the empty
ImmutableList), reset the property to that prototype value.