Moved List massaging to a method.
Since I allow the prototype value to be an ImmutableList with
elements in it, we need to do the modifications prior to
any modification, not just adds.

This is cleaner anyway.

Note that if you start with a populated ImmutableList, make some edits,
but end up editing it back to what it was originally, the editor
does not convert the value back to an ImmutableList, however that
shouldn't matter because if it's equal() to the prototype value
then it won't be saved anyway!