Don't reuse editor components.
TENTATIVE.
Heads up! Big change to how the config editor works.
But in reality this should fix some of the problems we already have...

Reusing the components is a pointless optimization and has caused
many issues over the years. Rather than dig-in to a current problem
further, let's nuke it.