Suppress gross warnings.
What a fiasco.
So, during flattening, we're replacing configs in the config manager.
That causes the old one to fire updated(), and re-gathers references.
Well, that is bound to fail because half the configs are in the process
of being flattened at that point.

Suppress warnings in the Default gatherer.

Gah. What would be better is to suppress all these updates during config
remapping like that. It's just causing extra pointless work during
flattening.