Fix pasting...
Turns out, you can "natively" paste in a ConfigTree, and
it follows a completely different code path than the other kind of
paste.

One would think we should unite those, but I'm not positive because
the ConfigTree is used elsewhere.

In any case, it was easy enough, once tracked down, to set up a
PasteHelper and have the ConfigTree use the same one from the editor...